feat: remove syslog binaries for real devices
