+++ lib/commands/log.js
@@ -2,8 +2,8 @@ import _ from 'lodash';
 import { DEFAULT_WS_PATHNAME_PREFIX } from 'appium-base-driver';
 import { iosCommands } from 'appium-ios-driver';
 import { IOSCrashLog } from '../device-log/ios-crash-log';
-import { IOSSimulatorLog } from '../device-log/ios-log';
-import { IOSDeviceLog } from '../device-log/real-device-ios-log';
+import { IOSSimulatorLog } from '../device-log/ios-simulator-log';
+import { IOSDeviceLog } from '../device-log/ios-device-log';
 import log from '../logger';
 import WebSocket from 'ws';
 import SafariConsoleLog from '../device-log/safari-console-log';
