+++ lib/device-log/ios-log.js
@@ -1,38 +1,62 @@
-import { IOSLog as IOSDriverIOSLog } from 'appium-ios-driver';
-import _ from 'lodash';
-import log from '../logger';
-import { SubProcess, exec } from 'teen_process';
+import { EventEmitter } from 'events';
 
+// We keep only the most recent log entries to avoid out of memory error
+const MAX_LOG_ENTRIES_COUNT = 10000;
 
-class IOSSimulatorLog extends IOSDriverIOSLog {
-  async startCapture () {
-    if (_.isUndefined(this.sim.udid)) {
-      throw new Error(`Log capture requires a sim udid`);
-    }
+class IOSLog extends EventEmitter {
 
-    if (!await this.sim.isRunning()) {
-      throw new Error(`iOS Simulator with udid ${this.sim.udid} is not running`);
-    }
-    const tool = 'xcrun';
-    const args = ['simctl', 'spawn', this.sim.udid, 'log', 'stream', '--style', 'compact'];
-    log.debug(`Starting log capture for iOS Simulator with udid '${this.sim.udid}', ` +
-              `using '${tool} ${args.join(' ')}'`);
-    try {
-      // cleanup existing listeners if the previous session has not been terminated properly
-      await exec('pkill', ['-xf', [tool, ...args].join(' ')]);
-    } catch (ign) {}
-    try {
-      this.proc = new SubProcess(tool, args);
-      await this.finishStartingLogCapture();
-    } catch (e) {
-      throw new Error(`Simulator log capture failed. Original error: ${e.message}`);
-    }
+  constructor () {
+    super();
+    this.logs = [];
+    this.logRow = '';
+    this.logIdxSinceLastRequest = -1;
+    this.maxBufferSize = MAX_LOG_ENTRIES_COUNT;
+  }
+
+  async startCapture () { // eslint-disable-line require-await
+    throw new Error(`Sub-classes need to implement a 'startCapture' function`);
+  }
+
+  async stopCapture () { // eslint-disable-line require-await
+    throw new Error(`Sub-classes need to implement a 'stopCapture' function`);
   }
 
   get isCapturing () {
-    return !!(this.proc && this.proc.isRunning);
+    throw new Error(`Sub-classes need to implement a 'isCapturing' function`);
+  }
+
+  broadcast (logLine) {
+    const logObj = {
+      timestamp: Date.now(),
+      level: 'ALL',
+      message: logLine
+    };
+    this.logs.push(logObj);
+    this.emit('output', logObj);
+    if (this.logs.length > this.maxBufferSize) {
+      this.logs.shift();
+      if (this.logIdxSinceLastRequest > 0) {
+        --this.logIdxSinceLastRequest;
+      }
+    }
+  }
+
+  getLogs () {
+    if (this.logs.length && this.logIdxSinceLastRequest < this.logs.length) {
+      let result = this.logs;
+      if (this.logIdxSinceLastRequest > 0) {
+        result = result.slice(this.logIdxSinceLastRequest);
+      }
+      this.logIdxSinceLastRequest = this.logs.length;
+      return result;
+    }
+    return [];
+  }
+
+  getAllLogs () {
+    return this.logs;
   }
 }
 
-export { IOSSimulatorLog };
-export default IOSSimulatorLog;
+export { IOSLog };
+export default IOSLog;
